update remote method call to add_pubmed #4648
this works partially - it works for existing
Abstracts with the pubmed id. Creating new
Abstracts is failing now with a strange postgres error -
PG::NotNullViolation: ERROR:  null value in column "id" violates not-null constraint

This error is not restricted to just Abstracts but
the creation of most new records. I suspect this is
a local database issue from the restore of the pg_dump file.