Faculty publications can be imported from NUBIC faculty webservice.
The importer process can be configure to retrieve faculty publications
from either PubMed (as it currently does) or from the NUBIC faculty
webservice.

Also, we fixed the namespacing for the lib/latticegrid folder to be
lib/lattice_grid. This way the Rails autoloader will pick up files
within that directory properly.