Temporarily disable graph generation.
This is a temprorary fix until we can fix the graph generation
hanging rails-prod2. We noticed this problem with Feinberg
when generating the Feinberg School of Medicine graph and it usually
takes a little while to hang the server, but that can speed up if
multiple graphs are being generated.