Update gems for dev env
Probably should do this in more than one commit but most of this commit has to do with the updating of gems and the setting up the application on the development machine.

Obstacles:
rmagick - had to reinstall imagemagick (via home-brew) more than once and setup the particular libraries (brew link) then setup env variables
  e.g. C_INCLUDE_PATH=/usr/local/Cellar/imagemagick/6.8.9-7/include/ImageMagick-6/ gem install rmagick
pg - using Postgres.app we need to install the gem with configuration option with-pg-config
  e.g. gem install pg -- --with-pg-config=/Applications/Postgres.app/Contents/Versions/9.3/bin/pg_config

Also, had to revert back to ruby 1.9.3 (the same as on the server) but we should investigate what is used by LatticeGrid that doesn't work well with more modern Ruby releases.